requirements.txt
alembic==1.4.2
aniso8601==8.0.0
attrs==19.3.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
click==7.1.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.3
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.8.0
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
Mako==1.1.2
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.15
Werkzeug==1.0.0
wincertstore==0.2
zipp==3.1.0
